requirements.txt
alpaca-py
python-dotenv
matplotlib
textual
